Converting to ginkgo and moving to secret_controller_test
